requirements.txt
# Core dependencies
pandas>=1.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.0.0
seaborn>=0.11.0

# MLflow and tracking
mlflow>=2.15.1
dagshub>=0.1.0

# Project utilities
python-box==6.0.2
PyYAML>=6.0
tqdm>=4.64.0
ensure==1.0.2
joblib>=1.1.0
types-pyYAML>=6.0.0
python-dotenv>=1.0.0

# Web application
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
notebook>=6.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Install the package in development mode
-e .
